added compare ops
