+++ QueryableRest/src/QRest.Core/Operations/Boolean/EqualOperation.cs
@@ -1,30 +1,16 @@
-﻿using System.Collections.Generic;
-using System.Linq.Expressions;
+﻿using System.Linq.Expressions;
 
 namespace QRest.Core.Operations.Boolean
 {
-    public class EqualOperation : OperationBase
+    public class EqualOperation : CompareOperationBase
     {
         public override bool SupportsCall => true;
 
-        public override Expression CreateCallExpression(ParameterExpression root, Expression context, IReadOnlyList<Expression> arguments)
+        protected override Expression PickExpression(Expression a, Expression b)
         {
-            if (arguments.Count != 1)
-                throw new ExpressionCreationException();
-
-            var a = context;
-            var b = arguments[0];
-
-            if (a.Type != b.Type)
-            {
-                if (a.Type.IsAssignableFrom(b.Type))
-                    a = Expression.Convert(a, b.Type);
-                else if (b.Type.IsAssignableFrom(a.Type))
-                    b = Expression.Convert(b, a.Type);
-                else throw new ExpressionCreationException();
-            }
-
             return Expression.Equal(a, b);
         }
+
+
     }
 }
