+++ QueryableRest/src/QRest.Core/Operations/Boolean/NotEqualOperation.cs
@@ -1,24 +1,14 @@
-﻿using System.Collections.Generic;
-using System.Linq.Expressions;
+﻿using System.Linq.Expressions;
 
 namespace QRest.Core.Operations.Boolean
 {
-    public class NotEqualOperation : OperationBase
+    public class NotEqualOperation : CompareOperationBase
     {
         public override bool SupportsCall => true;
 
-        public override Expression CreateCallExpression(ParameterExpression root, Expression context, IReadOnlyList<Expression> arguments)
+        protected override Expression PickExpression(Expression a, Expression b)
         {
-            if (arguments.Count != 1)
-                throw new ExpressionCreationException();
-
-            var a = context;
-            var b = arguments[0];
-
-            if (a.Type != b.Type)
-                b = Expression.Convert(b, a.Type);
-
             return Expression.NotEqual(a, b);
-        } 
+        }
     }
 }
